requirements.txt
flask
opencv-python-headless
numpy
python-dotenv